refactor(label): move Label to its folder with theme interface
#443 #234
