+++ src/docs/pages/FormsPage.tsx
@@ -1,9 +1,10 @@
 import type { FC } from 'react';
 import { useState } from 'react';
 import { HiMail } from 'react-icons/hi';
-import { Button, Label, Radio, Select, Textarea, TextInput, ToggleSwitch } from '../../lib';
+import { Button, Radio, Select, Textarea, TextInput, ToggleSwitch } from '../../lib';
 import { Checkbox } from '../../lib/components/Checkbox';
 import { FileInput } from '../../lib/components/FileInput';
+import { Label } from '../../lib/components/Label';
 import type { CodeExample } from './DemoPage';
 import { DemoPage } from './DemoPage';
