+++ src/docs/pages/ModalPage.tsx
@@ -1,8 +1,9 @@
 import type { FC } from 'react';
 import { useState } from 'react';
 import { HiOutlineExclamationCircle } from 'react-icons/hi';
-import { Button, Label, Modal, Select, TextInput } from '../../lib';
+import { Button, Modal, Select, TextInput } from '../../lib';
 import { Checkbox } from '../../lib/components/Checkbox';
+import { Label } from '../../lib/components/Label';
 import type { CodeExample } from './DemoPage';
 import { DemoPage } from './DemoPage';
