+++ src/lib/components/FormControls/index.ts
@@ -1,5 +1,4 @@
 export * from './HelperText';
-export * from './Label';
 export * from './Radio';
 export * from './Select';
 export * from './Textarea';
