+++ src/lib/components/Label/Label.stories.tsx
@@ -3,7 +3,7 @@ import type { LabelProps } from './Label';
 import { Label } from './Label';
 
 export default {
-  title: 'Components/FormControls',
+  title: 'Components/Label',
   component: Label,
 } as Meta;
