+++ src/lib/components/Label/index.ts
